add specialized ClosableConcurrentStack for CancellableStack which doesn't need pop or removal
