+++ servicetalk-concurrent-api/src/main/java/io/servicetalk/concurrent/api/SingleProcessor.java
@@ -28,8 +28,8 @@ import javax.annotation.Nullable;
  * @param <T> The type of result of the {@link Single}.
  */
 final class SingleProcessor<T> extends Single<T> implements Processor<T, T> {
-    private final ClosableConcurrentStack<Subscriber<? super T>, Consumer<Subscriber<? super T>>> stack =
-            new ClosableConcurrentStack<>();
+    private final RelaxedClosableConcurrentStack<Subscriber<? super T>, Consumer<Subscriber<? super T>>> stack =
+            new RelaxedClosableConcurrentStack<>();
 
     @Override
     protected void handleSubscribe(final Subscriber<? super T> subscriber) {
